+++ bindings/R/PPTree/src/main.cpp
@@ -10,34 +10,34 @@ using namespace RcppEigen;
 using namespace pptree;
 
 // [[Rcpp::export]]
-Tree<long double, int> pptree_train_glda(
-  const Data<long double> &data,
-  const DataColumn<int> &  groups,
-  const double             lambda) {
-  return Tree<long double, int>::train(
-    *TrainingSpec<long double, int>::glda(lambda),
-    DataSpec<long double, int>(data, groups));
+Tree<double, int> pptree_train_glda(
+  const Data<double> &    data,
+  const DataColumn<int> & groups,
+  const double            lambda) {
+  return Tree<double, int>::train(
+    *TrainingSpec<double, int>::glda(lambda),
+    DataSpec<double, int>(data, groups));
 }
 
 // [[Rcpp::export]]
-Forest<long double, int> pptree_train_forest_glda(
-  const Data<long double> & data,
-  const DataColumn<int> &   groups,
-  const int                 size,
-  const int                 n_vars,
-  const double              lambda,
-  SEXP                      n_threads) {
+Forest<double, int> pptree_train_forest_glda(
+  const Data<double> &    data,
+  const DataColumn<int> & groups,
+  const int               size,
+  const int               n_vars,
+  const double            lambda,
+  SEXP                    n_threads) {
   if (n_threads == R_NilValue) {
-    return Forest<long double, int>::train(
-      *TrainingSpec<long double, int>::uniform_glda(n_vars, lambda),
-      DataSpec<long double, int>(data, groups),
+    return Forest<double, int>::train(
+      *TrainingSpec<double, int>::uniform_glda(n_vars, lambda),
+      DataSpec<double, int>(data, groups),
       size,
       R::runif(0, INT_MAX));
   }
 
-  return Forest<long double, int>::train(
-    *TrainingSpec<long double, int>::uniform_glda(n_vars, lambda),
-    DataSpec<long double, int>(data, groups),
+  return Forest<double, int>::train(
+    *TrainingSpec<double, int>::uniform_glda(n_vars, lambda),
+    DataSpec<double, int>(data, groups),
     size,
     R::runif(0, INT_MAX),
     as<const int>(n_threads));
@@ -45,32 +45,32 @@ Forest<long double, int> pptree_train_forest_glda(
 
 // [[Rcpp::export]]
 DataColumn<int> pptree_predict(
-  const Tree<long double, int> &tree,
-  const Data<long double> &     data) {
+  const Tree<double, int> &tree,
+  const Data<double> &     data) {
   return tree.predict(data);
 }
 
 // [[Rcpp::export]]
 DataColumn<int> pptree_predict_forest(
-  const Forest<long double, int> &forest,
-  const Data<long double> &       data) {
+  const Forest<double, int> &forest,
+  const Data<double> &       data) {
   return forest.predict(data);
 }
 
 // [[Rcpp::export]]
-Data<long double> pptree_variable_importance(
-  const Tree<long double, int> &tree) {
-  return tree.variable_importance(VIProjectorStrategy<long double, int>());
+Data<double> pptree_variable_importance(
+  const Tree<double, int> &tree) {
+  return tree.variable_importance(VIProjectorStrategy<double, int>());
 }
 
 // [[Rcpp::export]]
-Data<long double> pptree_forest_variable_importance(
-  const Forest<long double, int> &forest) {
-  DataColumn<long double> projector = forest.variable_importance(VIProjectorStrategy<long double, int>());
-  DataColumn<long double> projector_adjusted = forest.variable_importance(VIProjectorAdjustedStrategy<long double, int>());
-  DataColumn<long double> permutation = forest.variable_importance(VIPermutationStrategy<long double, int>());
+Data<double> pptree_forest_variable_importance(
+  const Forest<double, int> &forest) {
+  DataColumn<double> projector = forest.variable_importance(VIProjectorStrategy<double, int>());
+  DataColumn<double> projector_adjusted = forest.variable_importance(VIProjectorAdjustedStrategy<double, int>());
+  DataColumn<double> permutation = forest.variable_importance(VIPermutationStrategy<double, int>());
 
-  Data<long double> result = Data<long double>(projector.size(), 3);
+  Data<double> result = Data<double>(projector.size(), 3);
 
   result.col(0) = projector;
   result.col(1) = projector_adjusted;
@@ -79,11 +79,11 @@ Data<long double> pptree_forest_variable_importance(
   return result;
 }
 
-Data<long double> parse_confusion_matrix(const ConfusionMatrix &confusion_matrix) {
+Data<double> parse_confusion_matrix(const ConfusionMatrix &confusion_matrix) {
   Data<int> values = confusion_matrix.values;
-  Data<long double> result(values.rows() + 1, values.cols() + 1);
+  Data<double> result(values.rows() + 1, values.cols() + 1);
 
-  DataColumn<long double> class_errors = confusion_matrix.class_errors();
+  DataColumn<double> class_errors = confusion_matrix.class_errors();
 
   for (int i = 0; i < values.rows(); i++) {
     for (int j = 0; j < values.cols(); j++) {
@@ -101,13 +101,13 @@ Data<long double> parse_confusion_matrix(const ConfusionMatrix &confusion_matrix
 }
 
 // [[Rcpp::export]]
-Data<long double> pptree_confusion_matrix(
-  const Tree<long double, int> &tree) {
+Data<double> pptree_confusion_matrix(
+  const Tree<double, int> &tree) {
   return parse_confusion_matrix(tree.confusion_matrix(*tree.training_data));
 }
 
 // [[Rcpp::export]]
-Data<long double> pptree_forest_confusion_matrix(
-  const Forest<long double, int> &forest) {
+Data<double> pptree_forest_confusion_matrix(
+  const Forest<double, int> &forest) {
   return parse_confusion_matrix(forest.confusion_matrix());
 }
