+++ core/src/ConfusionMatrix.test.cpp
@@ -144,7 +144,7 @@ TEST(ConfusionMatrix, ErrorMin) {
   DataColumn<int> predictions(3);
   predictions << 0, 1, 2;
 
-  long double result = ConfusionMatrix(predictions, actual).error();
+  double result = ConfusionMatrix(predictions, actual).error();
 
   ASSERT_DOUBLE_EQ(0, result);
 }
@@ -156,7 +156,7 @@ TEST(ConfusionMatrix, ErrorMax) {
   DataColumn<int> predictions(3);
   predictions << 2, 0, 1;
 
-  long double result = ConfusionMatrix(predictions, actual).error();
+  double result = ConfusionMatrix(predictions, actual).error();
 
   ASSERT_DOUBLE_EQ(1, result);
 }
@@ -168,7 +168,7 @@ TEST(ConfusionMatrix, ErrorGeneric) {
   DataColumn<int> predictions(6);
   predictions << 0, 1, 2, 0, 1, 2;
 
-  long double result = ConfusionMatrix(predictions, actual).error();
+  double result = ConfusionMatrix(predictions, actual).error();
 
   ASSERT_NEAR(0.5, result, 0.0001);
 }
@@ -180,9 +180,9 @@ TEST(ConfusionMatrix, ClassErrorsAllZero) {
   DataColumn<int> predictions(3);
   predictions << 0, 1, 2;
 
-  DataColumn<long double> result = ConfusionMatrix(predictions, actual).class_errors();
+  DataColumn<double> result = ConfusionMatrix(predictions, actual).class_errors();
 
-  DataColumn<long double> expected_errors(3);
+  DataColumn<double> expected_errors(3);
   expected_errors << 0, 0, 0;
 
   ASSERT_EQ(expected_errors.size(), result.size());
@@ -196,9 +196,9 @@ TEST(ConfusionMatrix, ClassErrorsAllOne) {
   DataColumn<int> predictions(6);
   predictions << 1, 2, 2, 1, 1, 1;
 
-  DataColumn<long double> result = ConfusionMatrix(predictions, actual).class_errors();
+  DataColumn<double> result = ConfusionMatrix(predictions, actual).class_errors();
 
-  DataColumn<long double> expected_errors(3);
+  DataColumn<double> expected_errors(3);
   expected_errors << 1, 1, 1;
 
   ASSERT_EQ(expected_errors.size(), result.size());
@@ -212,9 +212,9 @@ TEST(ConfusionMatrix, ClassErrorsMixed) {
   DataColumn<int> predictions(6);
   predictions << 0, 1, 2, 0, 1, 2;
 
-  DataColumn<long double> result = ConfusionMatrix(predictions, actual).class_errors();
+  DataColumn<double> result = ConfusionMatrix(predictions, actual).class_errors();
 
-  DataColumn<long double> expected_errors(3);
+  DataColumn<double> expected_errors(3);
   expected_errors << 0, 0.5, 0.666667;
 
   ASSERT_EQ(expected_errors.size(), result.size());
